PowerUp system for the player :
Add the bonuses implementation system, now by destroying hazards when the player points reach a certain amount, a bonus appear on the screen.
By touching this bonus, the player is provided with a specific ability, for example the double shot feature.
Some colisions bugs are now fixed, useless code has been deleted and the visual appearance of the bonus is better now.